LSTM Heroku Deploy
Deploying the LSTM Combined_Data using Flask/Heroku. It uses predicts the Closed Price of 21st day using past 20 days Close Price. In this we combined the News Articles data and Historical data to form new dataset named stock_data. We use this stock_data to train our model of Neural Networks which is build by using LSTM and Dense layers. At last we save the model in .h5 format, which is used by app.py file to display the results in the interface, index.html stored in templates folder. Finally the user can interact with the index.html file to enter a date for which he/se wants the Closed Price to be Predicted. We deploy this model using Flask/Heroku.